refactor: fetcher 함수명 수정
