+++ src/apis/alarm/updatePassword.ts
